making it compile on centos (gcc 4.1.1)
git-svn-id: https://devel.codership.com/svn/galera/trunk@946 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ galerautils/src/gu_time.h
@@ -24,8 +24,8 @@ gu_timeval_diff (struct timeval* left, struct timeval* right)
 static inline void
 gu_timeval_add (struct timeval* time, double s)
 {
-    long sec  = s;
-    long usec = (s - sec) * 1000000;
+    long sec  = (long)s;
+    long usec = (long)((s - sec) * 1000000);
     long carry;
     time->tv_usec += usec;
     carry = (time->tv_usec >= 1000000);
